compile warning on FreeBSD 4.x fixed (reported by Jesus Rodriguez)
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@337 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/lcr/lcr_mod.c
@@ -657,8 +657,8 @@ void print_gws (FILE *reply_file)
 		}
 		prefix_len = (*gws)[i].prefix_len;
 		if (prefix_len) {
-		    fprintf(reply_file, "%.*s\n",
-			    prefix_len, (*gws)[i].prefix);
+			fprintf(reply_file, "%.*s\n",
+				(int)prefix_len, (*gws)[i].prefix);
 		} else {
 		    fprintf(reply_file, "\n");
 		}
